// apps/backend/src/__generated__/users.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: 3mASDSO+VYF49YiDHz3L+RA0PdYZr/fvKmPgIMdOMRs3pRGdamRn37OBCPA6p5VtKTH/40h5EWX+pT/00ciQEw==
 */

/* eslint-disable */
// tslint:disable

interface Users {
  birthday: number
  created_on: number
  email: (string) | null
  /**
   * @default nextval('users_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'users_id'}
  password_hash: string
  profile_picture: (string) | null
  username: (string) | null
}
export default Users;

interface Users_InsertParameters {
  birthday: number
  created_on: number
  email?: (string) | null
  /**
   * @default nextval('users_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'users_id'}
  password_hash: string
  profile_picture?: (string) | null
  username?: (string) | null
}
export type {Users_InsertParameters}